Add Support for Kerberized Hive 3 (#257)
- Test against a static kerberized ambari server running HDFS 3.1.1 and Hive 3.1.0
- Add groups: multiClusterSecurity, proxySecurity for finer grain test grouping
- Remove `test_pxf_hdp2_secure_no_impersonation` and `test_pxf_hdp2_secure_with_impersonation` jobs from the pipeline as they were running a very basic test with Kerberos. These jobs are now replaced with `test_pxf_dataproc_multinode_with_impersonation` and `test_pxf_dataproc_multinode_no_impersonation`
- Add __UUID__ to the workingDirectory for hdfs

+++ concourse/scripts/pxf_common.bash
@@ -325,8 +325,10 @@ function init_and_configure_pxf_server() {
 		# Copy pxf-site.xml to the server configuration and update the
 		# pxf.service.user.name property value to use the PROXY_USER
 		# Only copy this file when testing against non-cloud
-		cp ${PXF_CONF_DIR}/templates/pxf-site.xml ${PXF_CONF_DIR}/servers/default/pxf-site.xml
-		sed -i -e "s|\${user.name}|${PROXY_USER}|g" ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+		if [[ ! -f ${PXF_CONF_DIR}/servers/default/pxf-site.xml ]]; then
+			cp ${PXF_CONF_DIR}/templates/pxf-site.xml ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+			sed -i -e "s|\${user.name}|${PROXY_USER}|g" ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+		fi
 	fi
 
 	# update runtime JDK value based on CI parameter
@@ -486,15 +488,51 @@ function configure_pxf_wasbs_server() {
 }
 
 function configure_pxf_default_server() {
-	# copy hadoop config files to PXF_CONF_DIR/servers/default
-	if [[ -d /etc/hadoop/conf/ ]]; then
-		cp /etc/hadoop/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
-	fi
-	if [[ -d /etc/hive/conf/ ]]; then
-		cp /etc/hive/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
-	fi
-	if [[ -d /etc/hbase/conf/ ]]; then
-		cp /etc/hbase/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
+	AMBARI_DIR=$(find /tmp/build/ -name ambari_env_files)
+	if [[ -n $AMBARI_DIR  ]]; then
+	  AMBARI_KEYTAB_FILE=$(find "$AMBARI_DIR" -name "*.keytab")
+		cp "${AMBARI_DIR}"/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
+
+		if [[ -n $AMBARI_KEYTAB_FILE ]]; then
+			REALM=$(cat "$AMBARI_DIR"/REALM)
+			HADOOP_USER=$(cat "$AMBARI_DIR"/HADOOP_USER)
+			cp ${PXF_CONF_DIR}/templates/mapred-site.xml ${PXF_CONF_DIR}/servers/default/mapred1-site.xml
+			cp ${PXF_CONF_DIR}/templates/pxf-site.xml ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+			sed -i -e "s|gpadmin/_HOST@EXAMPLE.COM|${HADOOP_USER}@${REALM}|g" ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+			sed -i -e "s|\${pxf.conf}/keytabs/pxf.service.keytab|$AMBARI_KEYTAB_FILE|g" ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+			sed -i -e "s|\${user.name}||g" ${PXF_CONF_DIR}/servers/default/pxf-site.xml
+			sudo mkdir -p /etc/security/keytabs/
+			sudo cp "$AMBARI_KEYTAB_FILE" /etc/security/keytabs/"${HADOOP_USER}".headless.keytab
+			sudo chown gpadmin:gpadmin /etc/security/keytabs/"${HADOOP_USER}".headless.keytab
+
+			mkdir -p ${PXF_CONF_DIR}/servers/db-hive/
+			cp ${PXF_CONF_DIR}/servers/default/pxf-site.xml ${PXF_CONF_DIR}/servers/db-hive/
+			cp ${PXF_CONF_DIR}/templates/jdbc-site.xml ${PXF_CONF_DIR}/servers/db-hive/
+
+			REALM=$(cat "$AMBARI_DIR"/REALM)
+			HIVE_HOSTNAME=$(grep < "$AMBARI_DIR"/etc_hostfile ambari-2 | awk '{print $2}')
+			KERBERIZED_HADOOP_URI="hive/${HIVE_HOSTNAME}.c.data-gpdb-ud.internal@${REALM};saslQop=auth" # quoted because of semicolon
+			sed -i -e 's|YOUR_DATABASE_JDBC_DRIVER_CLASS_NAME|org.apache.hive.jdbc.HiveDriver|' \
+				-e "s|YOUR_DATABASE_JDBC_URL|jdbc:hive2://${HIVE_HOSTNAME}:10000/default;principal=${KERBERIZED_HADOOP_URI}|" \
+				-e 's|YOUR_DATABASE_JDBC_USER||' \
+				-e 's|YOUR_DATABASE_JDBC_PASSWORD||' \
+				-e 's|</configuration>|<property><name>hadoop.security.authentication</name><value>kerberos</value></property></configuration>|g' \
+				${PXF_CONF_DIR}/servers/db-hive/jdbc-site.xml
+
+			PXF_SRC_DIR=$(find /tmp/build/ -name pxf_src)
+			cp "${PXF_SRC_DIR}"/automation/src/test/resources/hive-report.sql ${PXF_CONF_DIR}/servers/db-hive/
+		fi
+	else
+		# copy hadoop config files to PXF_CONF_DIR/servers/default
+		if [[ -d /etc/hadoop/conf/ ]]; then
+			cp /etc/hadoop/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
+		fi
+		if [[ -d /etc/hive/conf/ ]]; then
+			cp /etc/hive/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
+		fi
+		if [[ -d /etc/hbase/conf/ ]]; then
+			cp /etc/hbase/conf/*-site.xml "${PXF_CONF_DIR}/servers/default"
+		fi
 	fi
 }
